Support swappable schema model again.
Because we want to support a swappable schema model, we can't just duck-punch in a foreign
key to the admin.LogAdmin model, as it screws with ordering related to migrations.

Instead, we use a plain text column (and a property that does the fetch of the actual
schema model object, if necessary).

We also needed to move some migration operations around, so we can have the requirement
that our ('boardinghouse', '0001_initial') migration MUST run before the swapped-in
migration.

The example project uses a swapped model.

--HG--
branch : fix-swappable-model